(refactor|COS-113): Migrated auth to app router
