+++ packages/apps/spaces/app/auth/page.tsx
@@ -1,0 +1,3 @@
+export default async function Page() {
+  return <>Auth</>;
+}
