+++ packages/apps/spaces/app/auth/signin/CustomerOS-logo.png
