+++ packages/apps/spaces/app/auth/signin/login-bg.png
