+++ packages/apps/spaces/app/organizations/layout.tsx
@@ -1,0 +1,9 @@
+import { PageLayout } from '@shared/components/PageLayout';
+
+export default async function OrganizationsLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <PageLayout>{children}</PageLayout>;
+}
