+++ packages/apps/spaces/ui/media/logos/Google.tsx
@@ -1,0 +1,22 @@
+import { Icon, IconProps } from '@ui/media/Icon';
+
+export const Google = (props: IconProps) => (
+  <Icon viewBox='0 0 32 32' fill='none' boxSize='4' {...props}>
+    <path
+      d='M30.0014 16.3109C30.0014 15.1598 29.9061 14.3198 29.6998 13.4487H16.2871V18.6442H24.1601C24.0014 19.9354 23.1442 21.8798 21.2394 23.1864L21.2127 23.3604L25.4536 26.58L25.7474 26.6087C28.4458 24.1665 30.0014 20.5731 30.0014 16.3109Z'
+      fill='#4285F4'
+    />
+    <path
+      d='M16.2863 30C20.1434 30 23.3814 28.7555 25.7466 26.6089L21.2386 23.1865C20.0323 24.011 18.4132 24.5866 16.2863 24.5866C12.5086 24.5866 9.30225 22.1444 8.15929 18.7688L7.99176 18.7827L3.58208 22.1272L3.52441 22.2843C5.87359 26.8577 10.699 30 16.2863 30Z'
+      fill='#34A853'
+    />
+    <path
+      d='M8.15964 18.7688C7.85806 17.8977 7.68352 16.9643 7.68352 15.9999C7.68352 15.0354 7.85806 14.1021 8.14377 13.231L8.13578 13.0455L3.67083 9.64734L3.52475 9.71544C2.55654 11.6132 2.00098 13.7444 2.00098 15.9999C2.00098 18.2555 2.55654 20.3865 3.52475 22.2843L8.15964 18.7688Z'
+      fill='#FBBC05'
+    />
+    <path
+      d='M16.2864 7.4133C18.9689 7.4133 20.7784 8.54885 21.8102 9.4978L25.8419 5.64C23.3658 3.38445 20.1435 2 16.2864 2C10.699 2 5.8736 5.1422 3.52441 9.71549L8.14345 13.2311C9.30229 9.85555 12.5086 7.4133 16.2864 7.4133Z'
+      fill='#EB4335'
+    />
+  </Icon>
+);
